chatAgent/requirements.txt
python-dotenv
langchain
langchain-core
langchain-classic
langchain-community
langchain-groq
groq
fastapi
uvicorn